Add performance regression comparison in CI (#2362)
This commit adds a CI job that runs the `benchcomp` tool on the
performance regression suite, comparing the HEAD of the pull request to
the branch that the PR targets.

The CI job fails if any performance benchmark regresses when run using
the HEAD version of Kani with respect to the 'base' branch. Regression
is defined in the config file:

* a 3x slowdown for benchmarks that take <1s
* a 2x slowdown for benchmarks that take <30s
* a 10% slowdown for other benchmarks

or if any performance benchmark fails with the HEAD version while
passing with the base.

This fixes #2277.