Show only relevant benchmarks in CI summary
This commit adds a new key to the benchcomp markdown table
visualization. If `delete_row_if_empty` is set to true on a table's
extra column, then for each benchmark, if the extra column evaluates to
an empty string, then the benchmark will not be displayed in the table.

The change to the CI configuration means that benchmarks whose
solver, symex, or total runtime changed by less than 10% will not be
displayed in the tables for those metrics. The number of VCCs and number
of program steps tables will also not contain benchmarks for which those
numbers didn't change, and the success table will now only show
benchmarks that are either newly passing or newly failing.